Refactored the entity directory
