+++ CoworkingSpaceReservation/src/main/java/org/example/model/entity/BaseEntity.java
@@ -1,0 +1,23 @@
+package org.example.model.entity;
+
+
+import lombok.Data;
+import lombok.experimental.SuperBuilder;
+
+import java.io.Serializable;
+
+@Data
+@SuperBuilder
+public abstract class BaseEntity<T extends Serializable> implements Identifiable<T> {
+    private T id;
+
+    @Override
+    public T getId() {
+        return id;
+    }
+
+    @Override
+    public void setId(T id) {
+        this.id = id;
+    }
+}
