+++ CoworkingSpaceReservation/src/main/java/org/example/model/entity/Identifiable.java
