--- CoworkingSpaceReservation/src/main/java/org/example/model/entity/UserEntity.java
@@ -1,12 +1,0 @@
-package org.example.model.entity;
-
-import lombok.*;
-import lombok.experimental.SuperBuilder;
-
-@Data
-@SuperBuilder
-@NoArgsConstructor
-public abstract class UserEntity {
-    private long id;
-    private String login;
-}
