+++ CoworkingSpaceReservation/src/main/java/org/example/model/entity/accunt/AdminEntity.java
@@ -1,12 +1,12 @@
-package org.example.model.entity;
+package org.example.model.entity.accunt;
 
 import lombok.Data;
 import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
+import org.example.model.entity.BaseEntity;
+
 @EqualsAndHashCode(callSuper = true)
 @Data
 @SuperBuilder
-@NoArgsConstructor
-public class CustomerEntity extends UserEntity {
+public class AdminEntity extends BaseEntity<Long> {
 }
